+++ core-ng/src/main/java/core/framework/module/SiteConfig.java
@@ -3,6 +3,8 @@ package core.framework.module;
 import core.framework.http.HTTPMethod;
 import core.framework.internal.module.Config;
 import core.framework.internal.module.ModuleContext;
+import core.framework.internal.web.http.IPv4Ranges;
+import core.framework.internal.web.site.AJAXErrorResponse;
 import core.framework.internal.web.site.StaticContentController;
 import core.framework.internal.web.site.StaticDirectoryController;
 import core.framework.internal.web.site.StaticFileController;
@@ -77,5 +79,11 @@ public class SiteConfig extends Config {
             context.httpServer.handler.interceptors.add(webSecurityInterceptor);
         }
         return new WebSecurityConfig(webSecurityInterceptor);
+    }
+
+    public void publishAPI(List<String> cidrs) {
+        logger.info("publish api definition, cidrs={}", cidrs);
+        context.apiController.registry.beanClasses.add(AJAXErrorResponse.class);   // publish default ajax error response
+        context.apiController.accessControl.allow = new IPv4Ranges(cidrs);
     }
 }